==> Machine Learning - Teo Calvo/Machine Learning - Teo Calvo/ETL/2) Pedidos.ipynb <==
import pandas as pd
import pandasql as ps

pathOrders = '../data_curso/olist_orders_dataset.csv'
pathPedido = '../data/pedido.csv'

pedido = pd.read_csv(pathOrders)

pedido.head()

pedido = pedido.rename(columns={'order_id': 'idPedido', 'customer_id': 'idCliente', 'order_status': 'descSituacao',
                                'order_purchase_timestamp': 'dtPedido', 'order_approved_at': 'dtAprovado',
                                'order_delivered_carrier_date': 'dtEnvio', 'order_delivered_customer_date': 'dtEntrega',
                                'order_estimated_delivery_date': 'dtEntregaEstimada'})

pedido.head()

# pedido.to_csv(pathPedido, index = False, encoding = 'utf-8')

pedido[['dtPedido', 'dtAprovado', 'dtEnvio', 'dtEntrega', 'dtEntregaEstimada']].head()

start_date = '2017-07-01'
end_date = '2018-01-01'

# Filtrar o DataFrame para o período específico
filtered_df = pedido[(pedido['dtPedido'] >= start_date) & (pedido['dtPedido'] <= end_date)]

# Converter as colunas datetime para date
for col in filtered_df.columns:
    if pd.api.types.is_datetime64_any_dtype(filtered_df[col]):
        filtered_df[col] = filtered_df[col].dt.date

filtered_df.head()

dtPedido = pedido.loc[(pedido['dtPedido'] > '2017-07-01') & (pedido['dtPedido'] < '2018-01-01')]

datas = ['dtPedido', 'dtAprovado', 'dtEnvio', 'dtEntrega', 'dtEntregaEstimada']
dtPedido.loc[:, datas] = dtPedido.loc[:, datas].apply(lambda x: pd.to_datetime(x).dt.date)

dtPedido.head()

pathDtPedido = '../data/dt_pedido.csv'

# dtPedido.to_csv(pathDtPedido, index = False, encoding = 'utf-8')
