==> Machine Learning - Teo Calvo/Machine Learning - Teo Calvo/ETL/3) Item.ipynb <==
import pandas as pd

pathOrderItems = '../data_curso/olist_order_items_dataset.csv'
pathItem = '../data/item.csv'

item = pd.read_csv(pathOrderItems)

item.head()

item = item.rename(columns={'order_id': 'idPedido', 'order_item_id': 'idPedidoItem', 'product_id': 'idProduto',
                                        'seller_id': 'idVendedor', 'shipping_limit_date': 'shipping_limit_date',
                                        'price': 'vlPreco', 'freight_value': 'vlFrete'})

item.head()

#item.to_csv(pathItem, index = False, encoding = 'utf-8')

## Estudo SQL
usando sql no pandas

#!pip install -U pandasql

import pandasql as ps

pathItem = '../data/item.csv'

item = pd.read_csv(pathItem)

ps.sqldf("select * from item")

pathPedido = '../data/pedido.csv'

pedido = pd.read_csv(pathPedido)

tb_pedidos = ps.sqldf("""
                        select distinct
                            t1.idpedido,
                            t2.idVendedor
                            
                        from pedido as t1
                        
                        left join item as t2
                        on t1.idpedido = t2.idPedido                        
                        
                        where t1.dtPedido < '2018-01-01'
                        and t1.dtPedido >= '2017-07-01'
                        and t2.idVendedor is not null
                """)

tb_pedidos.head()

pathPagamento = '../data/pagamento.csv'

pagamento = pd.read_csv(pathPagamento)

tb_join = ps.sqldf("""select 
                            t1.idVendedor,
                            t2.*
                            
                        from tb_pedidos as t1
                        
                        left join pagamento as t2
                        on t1.idPedido = t2.idPedido
            """)

tb_join.head()

tb_group = ps.sqldf("""select 
                            idVendedor,
                            descTipoPagamento,
                            count(distinct idPedido) as qtdePedidoMeiopagamento,
                            sum(vlPagamento) vlPedidoMeiopagamento
                        from tb_join
                        group by idVendedor, descTipoPagamento
                        """
                    )

ps.sqldf("""select distinct(descTipoPagamento)
            from tb_group
            """
        )

vendedorQtdePagamentoTipo = ps.sqldf("""Select idVendedor,

                                    sum(case when descTipoPagamento = 'boleto' then qtdePedidoMeiopagamento else 0 end) as qtde_boleto,
                                    sum(case when descTipoPagamento = 'credit_card' then qtdePedidoMeiopagamento else 0 end) as qtde_credit_card,
                                    sum(case when descTipoPagamento = 'debit_card' then qtdePedidoMeiopagamento else 0 end) as qtde_debit_card,
                                    sum(case when descTipoPagamento = 'voucher' then qtdePedidoMeiopagamento else 0 end) as qtde_voucher,
                                    
                                    sum(case when descTipoPagamento = 'boleto' then vlPedidoMeiopagamento else 0 end) as valor_boleto,
                                    sum(case when descTipoPagamento = 'credit_card' then vlPedidoMeiopagamento else 0 end) as valor_credit_card,
                                    sum(case when descTipoPagamento = 'debit_card' then vlPedidoMeiopagamento else 0 end) as valor_debit_card,
                                    sum(case when descTipoPagamento = 'voucher' then vlPedidoMeiopagamento else 0 end) as valor_voucher,
                                    
                                    sum(case when descTipoPagamento = 'boleto' then qtdePedidoMeiopagamento else 0 end) / sum(qtdePedidoMeiopagamento) as proporcao_qtde_boleto,
                                    sum(case when descTipoPagamento = 'credit_card' then qtdePedidoMeiopagamento else 0 end) / sum(qtdePedidoMeiopagamento) as proporcao_qtde_credit_card,
                                    sum(case when descTipoPagamento = 'debit_card' then qtdePedidoMeiopagamento else 0 end) / sum(qtdePedidoMeiopagamento) as proporcao_qtde_debit_card,
                                    sum(case when descTipoPagamento = 'voucher' then qtdePedidoMeiopagamento else 0 end) / sum(qtdePedidoMeiopagamento) as proporcao_qtde_voucher,
                                    
                                    sum(case when descTipoPagamento = 'boleto' then vlPedidoMeiopagamento else 0 end) / sum(vlPedidoMeiopagamento) as proporcao_valor_boleto,
                                    sum(case when descTipoPagamento = 'credit_card' then vlPedidoMeiopagamento else 0 end) / sum(vlPedidoMeiopagamento) as proporcao_valor_credit_card,
                                    sum(case when descTipoPagamento = 'debit_card' then vlPedidoMeiopagamento else 0 end) / sum(vlPedidoMeiopagamento) as proporcao_valor_debit_card,
                                    sum(case when descTipoPagamento = 'voucher' then vlPedidoMeiopagamento else 0 end) / sum(vlPedidoMeiopagamento) as proporcao_valor_voucher
                                    
                                    from tb_group
                                    group by idVendedor
                                    """)
                                    

vendedorQtdePagamentoTipo.head()

tb_cartao = ps.sqldf("""
                        select idVendedor,
                        avg(nrParcelas) as avgQtdParcelas,
                        max(nrParcelas) as MaxQtdParcelas,
                        min(nrParcelas) as MinQtdParcelas
                        
                        from tb_join
                        
                        where descTipoPagamento = 'credit_card'
                        
                        group by idVendedor

                """)

tb_cartao.head()

tb_summary_cartao = ps.sqldf("""
                        select 
                        '2018-01-01' as dtReference,
                        t1.*,
                        t2.avgQtdParcelas,
                        t2.MaxQtdParcelas,
                        t2.MinQtdParcelas
                        
                        from vendedorQtdePagamentoTipo as t1
                        
                        left join tb_cartao as t2
                        on t1.idVendedor = t2.idVendedor
                """)

tb_summary_cartao.head()

pathVendedorPagamento = '../data/fs_vendedor_pagamento.csv'

# tb_summary_cartao.to_csv(pathVendedorPagamento, index = False, encoding = 'utf-8')
