==> Machine Learning - Teo Calvo/Machine Learning - Teo Calvo/ETL/4) Cliente.ipynb <==
import pandas as pd
import pandasql as ps

path_olist_customers_dataset = '../data_curso/olist_customers_dataset.csv'
pathCliente = '../data/cliente.csv'

cliente = pd.read_csv(path_olist_customers_dataset)

cliente.head()

cliente = cliente.rename(columns={'customer_id': 'idCliente', 
                                  'customer_unique_id': 'idClienteUnico',
                                  'customer_zip_code_prefix': 'codCep',
                                  'customer_city': 'descCidade',
                                  'customer_state': 'descUF' })

cliente.head()

#cliente.to_csv(pathCliente, index = False, encoding = 'utf-8')

## Região vendida

pathPedido = '../data/pedido.csv'

pedido = pd.read_csv(pathPedido)

pedido.head()

pathItem = '../data/item.csv'

item = pd.read_csv(pathItem)

item.head()

tb_regiao_cliente = ps.sqldf("""
                                select distinct
                                t1.idPedido,
                                t1.idCliente,
                                t2.idVendedor,
                                t3.descUF
                                
                                from pedido as t1
                                
                                left join item as t2
                                on t1.idPedido = t2.idPedido
                                
                                left join cliente as t3
                                on t1.idCliente = t3.idCliente
                                
                                where t1.dtPedido < '2018-01-01'
                                and t1.dtPedido >= '2017-07-01'
                                and t2.idVendedor is not null
                        """)

tb_regiao_cliente.head()

tb_estados = ps.sqldf("""
                        select distinct
                        descUF
                        
                        from tb_regiao_cliente
                        
                        order by descUF
                """)

tb_estados.head()

pd.set_option('display.max_columns', None)

tb_group = ps.sqldf("""
                        select 
                        '2018-01-01' as dtReference,
                        idvendedor,
                        count(idvendedor) as total,
                        count(distinct descUF) as qtdEstados,
                        cast(count(distinct case when descUF = 'AC' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteAC,
                        cast(count(distinct case when descUF = 'AL' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteAL,
                        cast(count(distinct case when descUF = 'AM' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteAM,
                        cast(count(distinct case when descUF = 'AP' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteAP,
                        cast(count(distinct case when descUF = 'BA' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteBA,
                        cast(count(distinct case when descUF = 'CE' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteCE,
                        cast(count(distinct case when descUF = 'DF' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteDF,
                        cast(count(distinct case when descUF = 'ES' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteES,
                        cast(count(distinct case when descUF = 'GO' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteGO,
                        cast(count(distinct case when descUF = 'MA' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteMA,
                        cast(count(distinct case when descUF = 'MG' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteMG,
                        cast(count(distinct case when descUF = 'MS' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteMS,
                        cast(count(distinct case when descUF = 'MT' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteMT,
                        cast(count(distinct case when descUF = 'PA' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClientePA,
                        cast(count(distinct case when descUF = 'PB' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClientePB,
                        cast(count(distinct case when descUF = 'PE' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClientePE,
                        cast(count(distinct case when descUF = 'PI' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClientePI,
                        cast(count(distinct case when descUF = 'PR' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClientePR,
                        cast(count(distinct case when descUF = 'RJ' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteRJ,
                        cast(count(distinct case when descUF = 'RN' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteRN,
                        cast(count(distinct case when descUF = 'RO' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteRO,
                        cast(count(distinct case when descUF = 'RR' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteRR,
                        cast(count(distinct case when descUF = 'RS' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteRS,
                        cast(count(distinct case when descUF = 'SC' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteSC,
                        cast(count(distinct case when descUF = 'SE' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteSE,
                        cast(count(distinct case when descUF = 'SP' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteSP,
                        cast(count(distinct case when descUF = 'TO' then idPedido end) as double) / cast(count(distinct idPedido) as double) as pctClienteTO
                        
                        from tb_regiao_cliente
                        
                        group by idVendedor
                """)

tb_group.head()

pathVendedorCliente = '../data/fs_vendedor_cliente.csv'

tb_group.to_csv(pathVendedorCliente, index = False, encoding = 'utf-8')
