==> Machine Learning - Teo Calvo/Machine Learning - Teo Calvo/ETL/5) Produto.ipynb <==
import pandas as pd
import pandasql as ps

pathPedido = '../data/pedido.csv'
pathItemPedido = '../data/item_pedido.csv'

pedido = pd.read_csv(pathPedido)
itemPedido = pd.read_csv(pathItemPedido)

pathproducts = '../data_curso/olist_products_dataset.csv'
pathProduto = '../data/produto.csv'

produto = pd.read_csv(pathproducts)

produto.head()

produto = produto.rename(columns = {"product_id" : "idProduto", 
                                    "product_category_name" : "descCategoria",
                                    "product_name_lenght" : "nrTamanhoNome", 
                                    "product_description_lenght" : "nrTamanhoDescricao",
                                    "product_photos_qty" : "nrFotos",
                                    "product_weight_g" : "vlPesoGramas",
                                    "product_length_cm" : "vlComprimentoCm",
                                    "product_height_cm" : "vlAlturaCm",
                                    "product_width_cm" : "vlLarguraCm"
    
})

produto.head()

#produto.to_csv(pathProduto, index = False, encoding = 'utf-8')

tb_join = ps.sqldf("""
                        select distinct
                        t2.idVendedor,
                        t3.*
                        
                        from pedido as t1
                        
                        left join itemPedido as t2
                        on t1.idPedido = t2.idPedido
                        
                        left join produto as t3
                        on t2.idProduto = t3.idProduto
                        
                         WHERE t1.dtPedido < '2018-01-01'
                         and t1.dtPedido >= '2017-07-01'
                         and t2.idVendedor is not null
                """)

tb_join.head()

tb_summary = ps.sqldf("""
                        select idVendedor,
                        avg(coalesce(nrFotos, 0)) as avgFotos,
                        avg(coalesce(vlComprimentoCm * vlAlturaCm * vlLarguraCm, 0)) as avgVolumeProduto,
                        min(coalesce(vlComprimentoCm * vlAlturaCm * vlLarguraCm, 0)) as MinimoVolumeProduto,
                        max(coalesce(vlComprimentoCm * vlAlturaCm * vlLarguraCm, 0)) as MaximoVolumeProduto
                        
                        from tb_join
                        
                        group by idVendedor
                """)

tb_summary.head()

tb_categoria = ps.sqldf("""
                        select descCategoria
                        
                        from itemPedido as t2
                        
                        left join produto as t3
                        on t2.idProduto = t3.idProduto
                        
                        and t2.idVendedor is not null
                        
                        group by descCategoria
                        order by count(distinct idPedido) desc
                        limit 15
                """)

tb_categoria.head(20)

tb_categoria.shape

tb_summary = ps.sqldf("""
                        select 
                        '2018-01-01' as dtReference,
                        idVendedor,
                        avg(coalesce(nrFotos, 0)) as avgFotos,
                        avg(coalesce(vlComprimentoCm * vlAlturaCm * vlLarguraCm, 0)) as avgVolumeProduto,
                        min(coalesce(vlComprimentoCm * vlAlturaCm * vlLarguraCm, 0)) as MinimoVolumeProduto,
                        max(coalesce(vlComprimentoCm * vlAlturaCm * vlLarguraCm, 0)) as MaximoVolumeProduto,
                        
                        cast(count (distinct case when descCategoria = 'cama_mesa_banho' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_cama_mesa_banho,
                        cast(count (distinct case when descCategoria = 'beleza_saude' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_beleza_saude,
                        cast(count (distinct case when descCategoria = 'esporte_lazer' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_esporte_lazer,
                        cast(count (distinct case when descCategoria = 'informatica_acessorios' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_informatica_acessorios,
                        cast(count (distinct case when descCategoria = 'moveis_decoracao' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_moveis_decoracao,
                        cast(count (distinct case when descCategoria = 'utilidades_domesticas' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_utilidades_domesticas,
                        cast(count (distinct case when descCategoria = 'relogios_presentes' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_relogios_presentes,
                        cast(count (distinct case when descCategoria = 'telefonia' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_telefonia,
                        cast(count (distinct case when descCategoria = 'automotivo' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_automotivo,
                        cast(count (distinct case when descCategoria = 'brinquedos' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_brinquedos,
                        cast(count (distinct case when descCategoria = 'cool_stuff' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_cool_stuff,
                        cast(count (distinct case when descCategoria = 'ferramentas_jardim' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_ferramentas_jardim,
                        cast(count (distinct case when descCategoria = 'perfumaria' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_perfumaria,
                        cast(count (distinct case when descCategoria = 'bebes' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_bebes,
                        cast(count (distinct case when descCategoria = 'eletronicos' then idProduto end) as double) / cast(count(distinct idProduto) as double) as pct_eletronicos
                        
                        from tb_join
                        
                        group by idVendedor
                """)

tb_summary.head()

pathVendedorProduto = '../data/fs_vendedor_produto.csv'

tb_summary.to_csv(pathVendedorProduto, index = False, encoding = 'utf-8')
