==> Machine Learning - Teo Calvo/Machine Learning - Teo Calvo/ETL/6) Avaliacao.ipynb <==
import pandas as pd
import pandasql as ps

pathPedido = '../data/pedido.csv'
pathItem = '../data/item.csv'

pedido = pd.read_csv(pathPedido)
item = pd.read_csv(pathItem)

pedido.head()

itemPedido.head()

tb_pedidos = ps.sqldf("""
                        select distinct
                            t1.idpedido,
                            t2.idVendedor
                            
                        from pedido as t1
                        
                        left join item as t2
                        on t1.idpedido = t2.idPedido                        
                        
                        where t1.dtPedido < '2018-01-01'
                        and t1.dtPedido >= '2017-07-01'
                        and t2.idVendedor is not null
                """)

tb_pedidos.head()

path_order_reviews_dataset = '../data_curso/olist_order_reviews_dataset.csv'
pathAvaliacaoPedido = '../data/avaliacao.csv'

avaliacaoPedido = pd.read_csv(path_order_reviews_dataset)

avaliacaoPedido.head()

avaliacaoPedido = avaliacaoPedido.rename(columns = {'review_id' : 'idAvaliacao', 'order_id' : 'idPedido', 
                                                    'review_score' : 'vlNota', 'review_comment_title' : 'desTituloComentario',
                                                    'review_comment_message' : 'desMensagemComentario',
                                                    'review_creation_date' : 'dataAvaliacao', 
                                                    'review_answer_timestamp' : 'dataResposta'    
})

avaliacaoPedido.head()

#avaliacaoPedido.to_csv(pathAvaliacaoPedido, index = False, encoding = 'utf-8')

tb_join = ps.sqldf(""" select 
                            t1.*,
                            t2.vlNota
                        from tb_pedidos as t1
                        
                        left join avaliacaoPedido as t2
                        on t1.idPedido = t2.idPedido
                    """)

tb_join.head()

tb_summary = ps.sqldf(""" select 
                            '2018-01-01' as dtReference,
                            idVendedor,
                            avg(vlNota) as mediaNota,
                            min(vlNota) as minNota,
                            max(vlNota) as maxNota,
                            cast(count(vlNota) as double) / cast(count(idPedido) as double) as pctAvaliacao
                            
                        from tb_join
                        
                        group by idVendedor
                    """)

tb_summary.head(100)

pathVendedorAvaliacao = '../data/fs_vendedor_avaliacao.csv'

# tb_summary.to_csv(pathVendedorAvaliacao, index = False, encoding = 'utf-8')
