==> Machine Learning - Teo Calvo/Machine Learning - Teo Calvo/ETL/7) Entrega.ipynb <==
import pandas as pd
import pandasql as ps

pedido = pd.read_csv("../data/pedido.csv")
itemPedido = pd.read_csv("../data/item_pedido.csv")

pedido.head()

itemPedido.head()

tb_pedido = ps.sqldf("""
                        select 
                            t1.idpedido,
                            t2.idVendedor,
                            t1.descSituacao,
                            t1.dtPedido,
                            t1.dtAprovado,
                            t1.dtEntrega,
                            t1.dtEntregaEstimada,
                            sum(vlFrete) as totalFrete
                            
                        from pedido as t1
                        
                        left join itemPedido as t2
                        on t1.idpedido = t2.idPedido                        
                        
                        where t1.dtPedido < '2018-01-01'
                        and t1.dtPedido >= '2017-07-01'
                        and t2.idVendedor is not null
                        
                        group by t1.idpedido,
                            t2.idVendedor,
                            t1.descSituacao,
                            t1.dtPedido,
                            t1.dtAprovado,
                            t1.dtEntrega,
                            t1.dtEntregaEstimada
                """)

tb_pedido.head()

tb_summary = ps.sqldf(""" select 
                            '2018-01-01' as dtReference,
                            idVendedor,
                            cast(count(distinct case when date(coalesce( dtEntrega, 2018-01-01)) > date(dtEntregaEstimada)
                                then idPedido 
                            end) as double) / 
                            cast(count(distinct case when descSituacao = 'delivered' then idPedido end) as double) 
                                as pctPedidoAtraso,
                            cast(count(distinct case when descSituacao = 'canceled' then idPedido end) as double) /
                                cast( count(distinct idPedido) as double) as pctPedidoCancelado,
                                avg(totalFrete) as avgFrete,
                                max(totalFrete) as maxFrete,
                                min(totalFrete) as minFrete,
                                avg(coalesce(dtEntrega, '2018-01-01') - dtPedido) as mediaDiasEntrega,
                                avg(dtEntregaEstimada - coalesce(dtEntrega, '2018-01-01')) as qtdDiasEntregaPromessa
                                
                        from tb_pedido
                        
                        group by idVendedor
                            
                    
                """)

tb_summary.head(30)

pathVendedorEntrega = '../data/fs_vendedor_entrega.csv'

tb_summary.to_csv(pathVendedorEntrega, index = False, encoding = 'utf-8')
