==> Machine Learning - Teo Calvo/Machine Learning - Teo Calvo/ETL/8) Venda.ipynb <==
import pandas as pd
import pandasql as ps

pathPedido = '../data/pedido.csv'
pathItem = '../data/item.csv'

pedido = pd.read_csv(pathPedido)

pedido.head()

item = pd.read_csv(pathItem)

item.head()

tb_pedido_item = ps.sqldf(""" select
                                    t2.*,
                                    t1.dtPedido
                              from pedido as t1

                              left join item as t2
                              on t1.idPedido = t2.idPedido

                              WHERE t1.dtPedido < '2018-01-01'
                              and t1.dtPedido >= '2017-07-01'
                              and t2.idVendedor is not null
                """)

tb_pedido_item.head()

tb_summary = ps.sqldf("""
                    select idVendedor,
                    count(distinct idPedido) as qtdPedidos,
                    count(distinct date(dtPedido)) as qtdDias,
                    count(idProduto) as qtdItens,
                    (JulianDay('2018-01-01') - JulianDay(max(dtPedido))) as qtdRecencia,
                    sum(vlPreco) / count(distinct idPedido) as avgTicket,
                    avg(vlPreco) as avgValorProduto,
                    max(vlPreco) as maxValorProduto,
                    min(vlPreco) as minValorProduto,
                    count(idProduto) / count(distinct idPedido) as avgProdutoPedido
                    
                    from tb_pedido_item      
                    
                    group by idVendedor
""")

tb_summary.head()

tb_pedido_summary = ps.sqldf("""
                        select 
                            idVendedor,
                            idPedido,
                            sum(vlPreco) as vlPreco
                            
                        from tb_pedido_item
                        
                        group by idVendedor, idPedido
                            

                """)

tb_pedido_summary.head()

tb_min_max = ps.sqldf("""
                    select 
                        idVendedor,
                        max(vlPreco) as maxVlPedido,
                        min(vlPreco) as minVlPedido
                        
                    from tb_pedido_summary
                    
                    group by idVendedor
""")

tb_min_max.head()

tb_life = ps.sqldf(""" 
                              select
                                    t2.idVendedor,
                                    sum(vlPreco) as  LTV,
                                    ('2918-01-01' -max(dtPedido)) as qtdDiasBase
                              from pedido as t1

                              left join item as t2
                              on t1.idPedido = t2.idPedido

                              WHERE t1.dtPedido < '2018-01-01'
                              and t2.idVendedor is not null
                              
                              group by t2.idVendedor
                """)

tb_life.head()

tb_dtPedido = ps.sqldf("""
                        select distinct
                            idVendedor,
                            date(dtPedido) as dtPedido,
                            lag(dtPedido) over(partition by idVendedor order by dtPedido) as lag1
                        
                        from tb_pedido_item                       
                        
                        order by idVendedor, dtPedido
""")

tb_dtPedido.head()

tb_intervalo = ps.sqldf("""
                    select 
                        idVendedor,
                        avg(JulianDay(dtPedido) - (JulianDay(lag1))) as avgIntervaloVendas
                    
                    from tb_dtPedido
                    group by idVendedor
""")

tb_intervalo.head(30)

tb_teste = ps.sqldf("""
                        select
                            '2018-01-01' as dtReference,
                            t1.*,
                            t2.maxVlPedido,
                            t2.minVlPedido,
                            t3.LTV,
                            t3.qtdDiasBase,
                            t4.avgIntervaloVendas
                            
                            
                        from tb_summary as t1
                        
                        left join tb_min_max as t2
                        on t1.idVendedor = t2.idVendedor
                        
                        left join tb_life as t3
                        on t1.idVendedor = t3.idVendedor
                        
                        left join tb_intervalo as t4
                        on t1.idVendedor = t4.idVendedor
""")

tb_teste.head()

pathVendedorVenda = '../data/fs_vendedor_venda.csv'

# tb_teste.to_csv(pathVendedorVenda, index = False, encoding = 'utf-8')
